--- plugins/deno_blocks_api/mod.ts
@@ -1,1 +1,0 @@
-export * from "./deno_blocks_api.ts";
