+++ routes/api/projects/index.ts
@@ -1,0 +1,54 @@
+import { Handlers } from "$fresh/server.ts";
+import { getSessionId } from "deno_kv_oauth/mod.ts";
+import { createProject } from "#/lib/subhosting_api/mod.ts";
+import { denoBlocksKv } from "#/lib/resources/deno_blocks_kv.ts";
+
+const DEPLOY_ORG_ID = Deno.env.get("DEPLOY_ORG_ID")!;
+const DEPLOY_ACCESS_TOKEN = Deno.env.get("DEPLOY_ACCESS_TOKEN")!;
+
+export const handler: Handlers = {
+  async GET(request: Request) {
+    const sessionID = await getSessionId(request);
+    if (!sessionID) {
+      return new Response("Unauthorized", { status: 401 });
+    }
+
+    // List project from Kv that are owned by the user.
+    const userID = await denoBlocksKv.getUserIDBySessionID({
+      sessionID,
+    });
+    if (!userID) {
+      return new Response("Unauthorized", { status: 401 });
+    }
+
+    const projects = await denoBlocksKv.getProjectsByUserID({
+      id: userID,
+    });
+    if (!projects) {
+      return new Response("Unauthorized", { status: 401 });
+    }
+
+    return Response.json(projects);
+  },
+  async POST(request: Request) {
+    const sessionID = await getSessionId(request);
+    if (!sessionID) {
+      return new Response("Unauthorized", { status: 401 });
+    }
+
+    // Create a new project for the user.
+    const userID = await denoBlocksKv.getUserIDBySessionID({
+      sessionID,
+    });
+    if (!userID) {
+      return new Response("Unauthorized", { status: 401 });
+    }
+
+    const project = await createProject({
+      organizationID: DEPLOY_ORG_ID,
+      accessToken: DEPLOY_ACCESS_TOKEN,
+    });
+    await denoBlocksKv.addProject({ userID, project });
+    return Response.json(project);
+  },
+};
